refactor: Move DrvBuildEvent to separate file
